WIP huge refactoring
