+++ test/users_test.go
@@ -3,7 +3,6 @@ package test
 import (
 	"bytes"
 	"encoding/json"
-	"github.com/khyurri/speedlog/engine/users"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 	"net/http"
